feat: Update timeout and memory options in Felice chatbot
This commit updates the timeout duration in the feliceCore Process function, changing it from 10 seconds to 20 seconds. This allows the chatbot more time to process and respond to messages.

In addition, the commit adds a new option to the chatgptCoreWithOptions configuration in the main function of the chatbot. The chatgpt.WithMemoryN(5) option is added, setting the memory size to 5. This means that the chatbot will store and utilize the last 5 interactions to generate more context-aware responses.

These changes aim to improve the performance and contextual understanding of the Felice chatbot.