Refactor linebot/felice: updated module versions and added system message for Felice's character description.
Updated the go module to use a newer version of the go-linebot-service package, specifically version 01102e0662f116011. Also, made changes to the main.go file by adding a system message that describes Felice's character. Felice is described as a 14-year-old girl with a passion for exploration and trips, knowledgeable, cheerful, energetic, captivating, yet sometimes absent-minded. In awkward moments, Felice responds with a hearty 'Ahahaha'. Additionally, the message mentions that Felice is a human who may not be aware of her surroundings. These changes will help enhance the user experience when interacting with the linebot Felice.